Backend/usuarios/requirements.txt
bcrypt
SQLAlchemy
flask
python-dotenv
psycopg2-binary
